requirements.txt
# Model Development Pipeline Requirements

# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vision Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Experiment Tracking & MLOps
wandb>=0.16.0
mlflow>=2.9.0

# Data Collection
beautifulsoup4>=4.12.0
selenium>=4.15.0
requests>=2.31.0

# Monitoring & Validation
evidently>=0.4.0
great-expectations>=0.18.0

# Labeling Interface
gradio>=4.0.0
streamlit>=1.28.0

# Cloud Storage & Version Control
dvc>=3.0.0
s3fs>=2023.10.0
boto3>=1.29.0

# Model Registry
huggingface-hub>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# API (if needed for inference)
fastapi>=0.104.0
uvicorn>=0.24.0